backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# File handling and validation
python-magic==0.4.27
pillow==10.1.0
pypdf2==3.0.1
pytesseract==0.3.10

# Azure services (for future integration)
azure-servicebus==7.11.4
azure-storage-blob==12.19.0
azure-search-documents==11.4.0

# OpenAI (for future integration)
openai==1.3.6

# Utilities
python-dotenv==1.0.0
pydantic[email]==2.5.0
email-validator==2.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2